+++ content/file/file_test.go
@@ -3052,65 +3052,6 @@ func TestStore_Dir_OverwriteSymlinkAbs(t *testing.T) {
 	}
 }
 
-func TestStore_Dir_OverwriteSymlink_RemovalFailed(t *testing.T) {
-	// prepare test content
-	tempDir, err := filepath.EvalSymlinks(t.TempDir())
-	if err != nil {
-		t.Fatal("error calling filepath.EvalSymlinks(), error =", err)
-	}
-	dirName := "testdir"
-	dirPath := filepath.Join(tempDir, dirName)
-	if err := os.MkdirAll(dirPath, 0777); err != nil {
-		t.Fatal("error calling Mkdir(), error =", err)
-	}
-
-	content := []byte("hello world")
-	fileName := "test.txt"
-	filePath := filepath.Join(dirPath, fileName)
-	if err := os.WriteFile(filePath, content, 0666); err != nil {
-		t.Fatal("error calling WriteFile(), error =", err)
-	}
-	// create symlink to an absolute path
-	symlink := filepath.Join(dirPath, "test_symlink")
-	if err := os.Symlink(filePath, symlink); err != nil {
-		t.Fatal("error calling Symlink(), error =", err)
-	}
-	if err := os.Chmod(symlink, 0444); err != nil {
-		t.Fatal("error calling Chmod(), error =", err)
-	}
-
-	src, err := New(tempDir)
-	if err != nil {
-		t.Fatal("Store.New() error =", err)
-	}
-	defer src.Close()
-	ctx := context.Background()
-
-	// add dir
-	desc, err := src.Add(ctx, dirName, "", dirPath)
-	if err != nil {
-		t.Fatal("Store.Add() error =", err)
-	}
-	// pack a manifest
-	opts := oras.PackManifestOptions{
-		Layers: []ocispec.Descriptor{desc},
-	}
-	manifestDesc, err := oras.PackManifest(ctx, src, oras.PackManifestVersion1_1, "test/dir", opts)
-	if err != nil {
-		t.Fatal("oras.PackManifest() error =", err)
-	}
-
-	// create a new store from the same root, to test overwriting symlink
-	dst, err := New(tempDir)
-	if err != nil {
-		t.Fatal("Store.New() error =", err)
-	}
-	defer dst.Close()
-	if err := oras.CopyGraph(ctx, src, dst, manifestDesc, oras.DefaultCopyGraphOptions); err == nil {
-		t.Error("oras.CopyGraph() error = nil, wantErr = true")
-	}
-}
-
 func TestCopyGraph_MemoryToFile_FullCopy(t *testing.T) {
 	src := memory.New()
